Added ability to generate and store results from games parentProcess can be used to create an arbitrarily large number of stock classifier dictionaries and then run games using these. It stores them in a single table in a mysql database.
The next step is to comment out the SARSA updating and try a simpler,
genetic algorithm that only changes the classifier dictionaries after
games, and sometimes randomly mutates the weights.

More and better features are also required. Specifically we need to get
the firing cone of a spotted enemy down.