Rewrote most of the bot. Temporarily abandoned sarsa.
The main additions are:

A graph representing the map
Commands passed as command objects instead of lists
Assigning paths as waypoint lists instead of simple objectives
Evaluating paths based on waypoint lists instead
An enemy belief system based on probable weighting of locations and the
graph
Visualization of the graph at simulation runtime.

All of this can be found in the graph_bot folder. Leaving the other stuff
instact for now.

Next step is:
Add features abusing probability density of enemies
Improve density calculations, atm they are just 1/d^2 which is mediocre.
Most promising approach is to abuse knowledge of likely enemy pathing.: